requirements.txt
JSONBender==0.9.3
pymongo==3.7.2
python-slugify